bff/requirements.txt
grpcio==1.69.0
grpcio-tools==1.69.0
protobuf==5.29.3
Flask==3.1.0
requests
PyJWT
python-dotenv
requests